correcting the resize of distance matrix
When the function is called with two agruments, the third is initialized
to a 1x1 matrix. This matrix is then resize to the number of points.
The problem is that it was done locally in the scope of the if
statement. Therefore, in the subsequent for loop the matrix is still 1x1
and the code crashes because of a stack overflow.